Keep last used command on history top
Previously, the order of the history commands never change. This was a result of
simply calling `add-to-list' to add the last used command to the history list
because `add-to-list' does nothing if the command is already in the list. The
solution is to always delete the command from the history list before pushing
the new command so that the last used command is always on the top.