Add :tempfile attribute for nim-mode #71
To be same behavior as command line, add :tempfile nil option.

For example, if there is file named array.nim and it is following
content, nim compiler emits error because of using same file name
and type name, but quickrun could success to compile if there is
no :tempfile attribute.

    var a: array[0..2, int] = [1, 2, 3]
    for ary in a:
      echo ary