[Style] Indent calls to quickrun-add-command like functions
This patch adds a declaration to `quickrun-add-command` so that Emacs
indents it like `defun` in user code.  For example, a form like...

    (quickrun-add-command "foo/bar"
                          '(:command ...))

...now indents by default like...

    (quickrun-add-command "foo/bar"
        '(:command ...))

This is naturally more readable since we can think of
`quickrun-add-command` as a function-defining-function: it creates
logic to execute for the given command, i.e. the first argument.  This
patch does not change any existing code within the package since the
change is only meant to affect user code that relies on Quickrun.

Signed-off-by: Eric James Michael Ritz <ericjmritz@yandex.com>