Ensure detection of '.C' extension as a C++ file
A file with extenson '.C' must be recognized as a C++ file.
Also, a user who has customized `case-fold-search` as nil,
still should be able to recognize the file `foo.F90` as
a Fortran file.

* quickrun.el (quickrun--decide-file-type):
Ignore the global value of `case-fold-search` when matching a file extension.
First, try to match the file extension with `case-fold-search` nil.
If no match, then try to match with `case-fold-search` non-nil.